Extracted code to make it easier to read and able to use within multiple functions. Add task with ajax now works
