+++ public/javascripts/manage-tasks.js
@@ -1,10 +1,12 @@
+import {editTaskMovement} from './index.js';
+import {deleteTaskFunction} from './delete-tasks.js';
 window.addEventListener("load", (event) => {
 
-    let deleteTaskButtons = document.querySelectorAll('.delete-button');
-    let tasksDivs = document.querySelectorAll('.task-div');
-    let editTasksDivs = document.querySelectorAll('.edit-buttons-div');
-    let allTotalTasks = document.querySelector('.tasks_incomplete');
-    let sum = document.querySelector('.sum');
+    // let deleteTaskButtons = document.querySelectorAll('.delete-button');
+    // let tasksDivs = document.querySelectorAll('.task-div');
+    // let editTasksDivs = document.querySelectorAll('.edit-buttons-div');
+    // let allTotalTasks = document.querySelector('.tasks_incomplete');
+    // let sum = document.querySelector('.sum');
 
     let addTaskInput = document.querySelector('.add-task-input');
     let addTaskForm = document.querySelector('.select-add-task-form');
@@ -13,8 +15,6 @@ window.addEventListener("load", (event) => {
     let orderedList = document.querySelector('.ordered-list');
 
     let originalPlaceholder = addTaskInput.placeholder
-
-    let length = []
 
     addTaskForm.addEventListener('submit', (e) => {
         e.preventDefault();
@@ -72,7 +72,7 @@ window.addEventListener("load", (event) => {
                 // SET VALUES AND INNER TEXT
                 newTask.innerHTML = data.taskContent;
                 openEditButton.innerText= 'v';
-                deleteTaskButtons.value=data.id;
+                deleteTaskButton.value=data.id;
                 deleteTaskButton.innerText= 'Delete';
                 closeEditButtonsDiv.innerText= '^';
 
@@ -85,6 +85,13 @@ window.addEventListener("load", (event) => {
                 taskContainer.append(editButtonsDiv);
                 editButtonsDiv.append(deleteTaskButton);
                 editButtonsDiv.append(closeEditButtonsDiv);  
+
+                editTaskMovement();
+
+                let deleteTaskButtons = document.querySelectorAll('.delete-button');
+
+                deleteTaskFunction(deleteTaskButtons);
+
             })
         }    
     })
@@ -92,41 +99,41 @@ window.addEventListener("load", (event) => {
 
 
 
-    // DELETE TASKS USING AJAX, UPDATE DOM UPON DELETE
-    if(deleteTaskButtons){
-        console.log(deleteTaskButtons)
-        for (let i=0; i<deleteTaskButtons.length; i++){
-            deleteTaskButtons[i].addEventListener('click', () => {
+    // // DELETE TASKS USING AJAX, UPDATE DOM UPON DELETE
+    // if(deleteTaskButtons){
+    //     console.log(deleteTaskButtons)
+    //     for (let i=0; i<deleteTaskButtons.length; i++){
+    //         deleteTaskButtons[i].addEventListener('click', () => {
 
-            fetch(`/tasks/delete`, {
-                method: "DELETE",
-                headers: {
-                    "Content-Type": "application/json"
-                },
-                body: JSON.stringify({
-                taskId: deleteTaskButtons[i].value
-                })
-            })
-            .then(function(res) {
-                if (!res.ok) {
-                    throw Error(res.statusText); // handle any potential server errors
-                }
-                return res.json();
-            })
-            .then(function(data) {
-                // tasks[i].remove();
-                if(allTotalTasks.id === data.userListId || !allTotalTasks.id){
-                    sum.innerText-=1
-                }
+    //         fetch(`/tasks/delete`, {
+    //             method: "DELETE",
+    //             headers: {
+    //                 "Content-Type": "application/json"
+    //             },
+    //             body: JSON.stringify({
+    //             taskId: deleteTaskButtons[i].value
+    //             })
+    //         })
+    //         .then(function(res) {
+    //             if (!res.ok) {
+    //                 throw Error(res.statusText); // handle any potential server errors
+    //             }
+    //             return res.json();
+    //         })
+    //         .then(function(data) {
+    //             // tasks[i].remove();
+    //             if(allTotalTasks.id === data.userListId || !allTotalTasks.id){
+    //                 sum.innerText-=1
+    //             }
                 
-                tasksDivs[i].remove();
-                editTasksDivs[i].remove();
+    //             tasksDivs[i].remove();
+    //             editTasksDivs[i].remove();
               
-            })
-            .catch(function(error) {
-                console.log(error)
-            });
-        })};
-    };
+    //         })
+    //         .catch(function(error) {
+    //             console.log(error)
+    //         });
+    //     })};
+    // };
 });
