+++ views/main.pug
@@ -9,6 +9,7 @@ append head
   script(src="/javascripts/index.js" type="module" defer)
   script(src="/javascripts/manage-tasks.js" type="module" defer)
   script(src="/javascripts/manage-lists.js" type="module" defer)
+  script(src="/javascripts/delete-tasks.js" type="module" defer)
 
 block content
     div.navbar
